Split utilities into two classes
