+++ src/test/java/org/junitpioneer/jupiter/PioneerUtilsTests.java
@@ -14,7 +14,6 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import java.util.HashSet;
-import java.util.List;
 import java.util.Set;
 import java.util.function.BiConsumer;
 import java.util.function.BinaryOperator;
@@ -24,270 +23,9 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
-import org.junit.platform.engine.test.event.ExecutionEvent;
-import org.junit.platform.engine.test.event.ExecutionEventRecorder;
-import org.junitpioneer.AbstractPioneerTestEngineTests;
 
 @DisplayName("JUnitPioneer utilities")
 class PioneerUtilsTests {
-
-	@Nested
-	@DisplayName("for annotations")
-	class AnnotationUtilsTests extends AbstractPioneerTestEngineTests {
-
-		/*
-		 * Cases covered here:
-		 *  - directly present
-		 *  - present on enclosing class
-		 *
-		 * Cases covered by Jupiter's `AnnotationSupport::findRepeatableAnnotations`:
-		 *  - indirectly present (i.e. inherited from super class)
-		 *  - meta present (i.e. on another annotation)
-		 */
-
-		@Nested
-		@DisplayName("that are not repeatable")
-		class SimpleAnnotations {
-
-			@Nested
-			@DisplayName("and not stackable")
-			class StopOnFirst {
-
-				@BeforeEach
-				void enableStopOnFirst() {
-					FailExtension.STACKABLE.set(false);
-				}
-
-				@Test
-				void discoversMethodAnnotation() {
-					ExecutionEventRecorder eventRecorder = executeTests(UtilsAnnotationTestCases.FailTestCases.class,
-						"methodIsAnnotated");
-					assertFailedTestHasMessage(eventRecorder, "method");
-				}
-
-				@Test
-				void discoversRootClassAnnotation() {
-					ExecutionEventRecorder eventRecorder = executeTests(
-						UtilsAnnotationTestCases.FailTestCases.RootClassTestCases.class, "rootClassIsAnnotated");
-					assertFailedTestHasMessage(eventRecorder, "root class");
-				}
-
-				@Test
-				void discoversClassAnnotation() {
-					ExecutionEventRecorder eventRecorder = executeTests(
-						UtilsAnnotationTestCases.FailTestCases.NestedTestCases.class, "classIsAnnotated");
-					assertFailedTestHasMessage(eventRecorder, "nested class");
-				}
-
-				@Test
-				void discoversOuterClassAnnotation() {
-					ExecutionEventRecorder eventRecorder = executeTests(
-						UtilsAnnotationTestCases.FailTestCases.NestedTestCases.TwiceNestedTestCases.class,
-						"outerClassIsAnnotated");
-					assertFailedTestHasMessage(eventRecorder, "nested class");
-				}
-
-				@Test
-				void discoversOuterOuterClassAnnotation() {
-					ExecutionEventRecorder eventRecorder = executeTests(
-						UtilsAnnotationTestCases.FailTestCases.NestedTestCases.TwiceNestedTestCases.ThriceNestedTestCases.class,
-						"outerClassIsAnnotated");
-					assertFailedTestHasMessage(eventRecorder, "nested class");
-				}
-
-			}
-
-			@Nested
-			@DisplayName("but stackable")
-			class Stackable {
-
-				@BeforeEach
-				void enableStopOnFirst() {
-					FailExtension.STACKABLE.set(true);
-				}
-
-				@Test
-				void discoversMethodAnnotation() {
-					ExecutionEventRecorder eventRecorder = executeTests(UtilsAnnotationTestCases.FailTestCases.class,
-						"methodIsAnnotated");
-					assertFailedTestHasMessage(eventRecorder, "method", "root class");
-				}
-
-				@Test
-				void discoversRootClassAnnotation() {
-					ExecutionEventRecorder eventRecorder = executeTests(
-						UtilsAnnotationTestCases.FailTestCases.RootClassTestCases.class, "rootClassIsAnnotated");
-					assertFailedTestHasMessage(eventRecorder, "root class");
-				}
-
-				@Test
-				void discoversClassAnnotation() {
-					ExecutionEventRecorder eventRecorder = executeTests(
-						UtilsAnnotationTestCases.FailTestCases.NestedTestCases.class, "classIsAnnotated");
-					assertFailedTestHasMessage(eventRecorder, "nested class", "root class");
-				}
-
-				@Test
-				void discoversOuterClassAnnotation() {
-					ExecutionEventRecorder eventRecorder = executeTests(
-						UtilsAnnotationTestCases.FailTestCases.NestedTestCases.TwiceNestedTestCases.class,
-						"outerClassIsAnnotated");
-					assertFailedTestHasMessage(eventRecorder, "nested class", "root class");
-				}
-
-				@Test
-				void discoversOuterOuterClassAnnotation() {
-					ExecutionEventRecorder eventRecorder = executeTests(
-						UtilsAnnotationTestCases.FailTestCases.NestedTestCases.TwiceNestedTestCases.ThriceNestedTestCases.class,
-						"outerClassIsAnnotated");
-					assertFailedTestHasMessage(eventRecorder, "nested class", "root class");
-				}
-
-			}
-
-		}
-
-		@Nested
-		@DisplayName("that are repeatable")
-		class RepeatableAnnotations {
-
-			@Nested
-			@DisplayName("but not stackable")
-			class StopOnFirst {
-
-				@BeforeEach
-				void enableStopOnFirst() {
-					RepeatableFailExtension.STACKABLE.set(false);
-				}
-
-				@Test
-				void discoversMethodAnnotation() {
-					ExecutionEventRecorder eventRecorder = executeTests(
-						UtilsAnnotationTestCases.RepeatableFailTestCases.class, "methodIsAnnotated");
-					assertFailedTestHasMessage(eventRecorder, "method");
-				}
-
-				@Test
-				void discoversRootClassAnnotation() {
-					ExecutionEventRecorder eventRecorder = executeTests(
-						UtilsAnnotationTestCases.RepeatableFailTestCases.RootClassTestCases.class,
-						"rootClassIsAnnotated");
-					assertFailedTestHasMessage(eventRecorder, "root class");
-				}
-
-				@Test
-				void discoversRepeatedMethodAnnotation() {
-					ExecutionEventRecorder eventRecorder = executeTests(
-						UtilsAnnotationTestCases.RepeatableFailTestCases.class, "methodIsRepeatablyAnnotated");
-					assertFailedTestHasMessage(eventRecorder, "repeated", "annotation");
-				}
-
-				@Test
-				void discoversClassAnnotation() {
-					ExecutionEventRecorder eventRecorder = executeTests(
-						UtilsAnnotationTestCases.RepeatableFailTestCases.NestedTestCases.class, "classIsAnnotated");
-					assertFailedTestHasMessage(eventRecorder, "nested class");
-				}
-
-				@Test
-				void discoversOuterClassAnnotation() {
-					ExecutionEventRecorder eventRecorder = executeTests(
-						UtilsAnnotationTestCases.RepeatableFailTestCases.NestedTestCases.TwiceNestedTestCases.class,
-						"outerClassIsAnnotated");
-					assertFailedTestHasMessage(eventRecorder, "nested class");
-				}
-
-				@Test
-				void discoversOuterOuterClassAnnotation() {
-					ExecutionEventRecorder eventRecorder = executeTests(
-						UtilsAnnotationTestCases.RepeatableFailTestCases.NestedTestCases.TwiceNestedTestCases.ThriceNestedTestCases.class,
-						"outerClassIsAnnotated");
-					assertFailedTestHasMessage(eventRecorder, "nested class");
-				}
-
-				@Test
-				void discoversRepeatedOuterClassAnnotation() {
-					ExecutionEventRecorder eventRecorder = executeTests(
-						UtilsAnnotationTestCases.RepeatableFailTestCases.NestedRepeatableTestCases.class,
-						"outerClassIsRepeatablyAnnotatedAnnotated");
-					assertFailedTestHasMessage(eventRecorder, "repeated", "annotation");
-				}
-
-			}
-
-			@Nested
-			@DisplayName("and stackable")
-			class Stackable {
-
-				@BeforeEach
-				void enableStopOnFirst() {
-					RepeatableFailExtension.STACKABLE.set(true);
-				}
-
-				@Test
-				void discoversMethodAnnotation() {
-					ExecutionEventRecorder eventRecorder = executeTests(
-						UtilsAnnotationTestCases.RepeatableFailTestCases.class, "methodIsAnnotated");
-					assertFailedTestHasMessage(eventRecorder, "method", "root class");
-				}
-
-				@Test
-				void discoversRootClassAnnotation() {
-					ExecutionEventRecorder eventRecorder = executeTests(
-						UtilsAnnotationTestCases.RepeatableFailTestCases.RootClassTestCases.class,
-						"rootClassIsAnnotated");
-					assertFailedTestHasMessage(eventRecorder, "root class");
-				}
-
-				@Test
-				void discoversRepeatedMethodAnnotation() {
-					ExecutionEventRecorder eventRecorder = executeTests(
-						UtilsAnnotationTestCases.RepeatableFailTestCases.class, "methodIsRepeatablyAnnotated");
-					assertFailedTestHasMessage(eventRecorder, "repeated", "annotation", "root class");
-				}
-
-				@Test
-				void discoversClassAnnotation() {
-					ExecutionEventRecorder eventRecorder = executeTests(
-						UtilsAnnotationTestCases.RepeatableFailTestCases.NestedTestCases.class, "classIsAnnotated");
-					assertFailedTestHasMessage(eventRecorder, "nested class", "root class");
-				}
-
-				@Test
-				void discoversOuterClassAnnotation() {
-					ExecutionEventRecorder eventRecorder = executeTests(
-						UtilsAnnotationTestCases.RepeatableFailTestCases.NestedTestCases.TwiceNestedTestCases.class,
-						"outerClassIsAnnotated");
-					assertFailedTestHasMessage(eventRecorder, "nested class", "root class");
-				}
-
-				@Test
-				void discoversOuterOuterClassAnnotation() {
-					ExecutionEventRecorder eventRecorder = executeTests(
-						UtilsAnnotationTestCases.RepeatableFailTestCases.NestedTestCases.TwiceNestedTestCases.ThriceNestedTestCases.class,
-						"outerClassIsAnnotated");
-					assertFailedTestHasMessage(eventRecorder, "nested class", "root class");
-				}
-
-				@Test
-				void discoversRepeatedOuterClassAnnotation() {
-					ExecutionEventRecorder eventRecorder = executeTests(
-						UtilsAnnotationTestCases.RepeatableFailTestCases.NestedRepeatableTestCases.class,
-						"outerClassIsRepeatablyAnnotatedAnnotated");
-					assertFailedTestHasMessage(eventRecorder, "repeated", "annotation", "root class");
-				}
-
-			}
-
-		}
-
-		private void assertFailedTestHasMessage(ExecutionEventRecorder eventRecorder, String... messages) {
-			List<ExecutionEvent> failed = eventRecorder.getFailedTestFinishedEvents();
-			assertThat(failed).hasSize(1);
-			assertThat(getFirstFailuresThrowable(eventRecorder).getMessage()).isEqualTo(String.join(",", messages));
-		}
-
-	}
 
 	@Nested
 	@DisplayName("to distinct set")
